week2/day5/task_management_api/requirements.txt
flask==2.3.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0
flask-cors==4.0.0 